Drop bundler-audit checks
There is no real need to run bundle-audit in a library. Vulnerabilities
will only manifest in the services using the library. Even keeping the
library dependencies up to date will not (not even close) guarantee that
the services keep _their_ dependencies up to date.

Also the bundle-audit checks feels like a constant drag when doing
library work: We need to bump patches or list ignored CVEs instead of
doing the actual work that needs doing.