drop bundler-audit checks
There is no real need to run bundle-audit on this gem’s test dependencies. It is better to test on more recent versions, but they are a security concern here. However, they slow down actual work on a gem: you need to bump patches or list ignored CVEs instead of working on something useful.